Automatically update internal links
This commit introduces the following changes:
  - Automatically fix outdated links when saving a content translation
  - Extract `clean_content` functionality into its own function
  - Update referencing translations when updating a content translation
  - Allow marking a link as auto update in tinymce
  - Automatically update link text if `data-integreat-auto-update` is true
  - Insert icons for event and poi links

This commit does not optimize the url query used to find referencing urls, since we can't know the last public title and slug:
As it turns out, once one page translation gets set to draft, all previous versions also become draft.
This makes it impossible to figure out the last public version, which is required for the optimization to work.
It is probably still possible to implement it, but for now I am not sure whether that is worth it.